Migrate options page to popup settings view and improve UI/UX
