--- src/pages/Options/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Settings</title>
-  </head>
-
-  <body>
-    <a
-      target="_blank"
-      rel="noopener noreferrer"
-      href="https://github.com/dedoussis/icloud-hide-my-email-browser-extension"
-    >
-      <img
-        loading="lazy"
-        width="149"
-        height="149"
-        src="https://github.blog/wp-content/uploads/2008/12/forkme_right_darkblue_121621.png?resize=149%2C149"
-        style="position: absolute; top: 0; right: 0; border: 0"
-        class="attachment-full size-full"
-        alt="Fork me on GitHub"
-        data-recalc-dims="1"
-      />
-    </a>
-    <div id="app-container"></div>
-  </body>
-</html>
