+++ webpack.config.js
@@ -52,7 +52,6 @@ const options = {
     popup: path.join(__dirname, 'src', 'pages', 'Popup', 'index.tsx'),
     background: path.join(__dirname, 'src', 'pages', 'Background', 'index.ts'),
     contentScript: path.join(__dirname, 'src', 'pages', 'Content', 'index.ts'),
-    options: path.join(__dirname, 'src', 'pages', 'Options', 'index.tsx'),
     userguide: path.join(__dirname, 'src', 'pages', 'Userguide', 'index.tsx'),
   },
   chromeExtensionBoilerplate: {
@@ -172,12 +171,6 @@ const options = {
       template: path.join(__dirname, 'src', 'pages', 'Popup', 'index.html'),
       filename: 'popup.html',
       chunks: ['popup'],
-      cache: false,
-    }),
-    new HtmlWebpackPlugin({
-      template: path.join(__dirname, 'src', 'pages', 'Options', 'index.html'),
-      filename: 'options.html',
-      chunks: ['options'],
       cache: false,
     }),
     new HtmlWebpackPlugin({
